refactor: Extract service tests
Modify test to be more modular, readable, and flexible.

- Add new service test file to modularize service related tests.
- Move service-related tests to new service test file.
- Arrange tests in chronological order.
- Define the service class in the TestCase 'setUp' function.
